[MOD] rockbotic_custom: event_registration add fields
