+++ rockbotic_custom/__openerp__.py
@@ -52,6 +52,7 @@
         "views/account_invoice_view.xml",
         "views/res_company_view.xml",
         "views/account_view.xml",
+        "views/event_registration_view.xml",
         "report/rockbotic_custom_view.xml",
         "report/rockbotic_layout_view.xml",
         "report/rockbotic_reports.xml",
